+++ src/command/replicate/mod.rs
@@ -112,7 +112,7 @@ pub async fn run(options: &Options, config: &Configuration) -> Result<(), Error>
     let checker = ReplicationChecker::new(metadata_store.clone(), repositories.clone());
 
     let mut drain = None;
-    let mut sink: Box<dyn ActionSink + Send> = if options.dry_run {
+    let sink: Box<dyn ActionSink> = if options.dry_run {
         info!("Dry-run mode: no changes will be made to the storage");
         Box::new(DryRunSink)
     } else {
@@ -133,7 +133,7 @@ pub async fn run(options: &Options, config: &Configuration) -> Result<(), Error>
         ))
     };
 
-    check::check_namespaces(&metadata_store, &checker, sink.as_mut()).await?;
+    check::check_namespaces(&metadata_store, &checker, sink.as_ref()).await?;
     if let Some(drain) = &drain {
         drain.drain().await;
     }