+++ src/command/scrub/mod.rs
@@ -1,9 +1,11 @@
 pub mod action;
+pub mod categorize;
 pub mod check;
 mod command;
 mod error;
 pub mod executor;
-pub mod setup;
+pub mod validate;
+pub mod walk;
 
 pub use command::{Command, Options};
 pub use error::Error;